+++ db/migrate/20220901102229_add_column_to_alcool_profile.rb
@@ -1,16 +1,16 @@
 class AddColumnToAlcoolProfile < ActiveRecord::Migration[7.0]
   def change
-    add_column :alcool_profiles, :vineux, :integer, default: 0
-    add_column :alcool_profiles, :epicee, :integer, default: 0
-    add_column :alcool_profiles, :boise, :integer, default: 0
-    add_column :alcool_profiles, :animale, :integer, default: 0
-    add_column :alcool_profiles, :noix, :integer, default: 0
-    add_column :alcool_profiles, :sucre, :integer, default: 0
-    add_column :alcool_profiles, :fruite, :integer, default: 0
-    add_column :alcool_profiles, :floral, :integer, default: 0
-    add_column :alcool_profiles, :herbace, :integer, default: 0
-    add_column :alcool_profiles, :cereale, :integer, default: 0
-    add_column :alcool_profiles, :empyreumatique, :integer, default: 0
-    add_column :alcool_profiles, :tourbe, :integer, default: 0
+    add_column :alcool_profiles, :vineux, :float, default: 0
+    add_column :alcool_profiles, :epicee, :float, default: 0
+    add_column :alcool_profiles, :boise, :float, default: 0
+    add_column :alcool_profiles, :animale, :float, default: 0
+    add_column :alcool_profiles, :noix, :float, default: 0
+    add_column :alcool_profiles, :sucre, :float, default: 0
+    add_column :alcool_profiles, :fruite, :float, default: 0
+    add_column :alcool_profiles, :floral, :float, default: 0
+    add_column :alcool_profiles, :herbace, :float, default: 0
+    add_column :alcool_profiles, :cereale, :float, default: 0
+    add_column :alcool_profiles, :empyreumatique, :float, default: 0
+    add_column :alcool_profiles, :tourbe, :float, default: 0
   end
 end
